requirements.txt
aiobotocore==1.3.0
aiohttp==3.7.4.post0
aioitertools==0.7.1
arabic-reshaper==2.1.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
botocore==1.20.49
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2.10
django-ajax==0.3.0
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-icons==2.2.1
django-jchart==0.4.2
django-jquery==3.1.0
django-mathfilters==1.0.0
django-money==1.3.1
django-readonly-field==1.0.5
django-rest-framework==0.1.0
django-select2==7.7.0
django-settings-context-processor==0.2
django-tables2==2.3.4
django-taggit==1.3.0
django-taggit-serializer==0.1.7
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.2
docopt==0.6.2
fanstatic==1.1
fsspec==0.9.0
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
importlib-metadata==2.1.1
jmespath==0.10.0
js.jquery==3.3.1
jsonfield==2.1.1
multidict==5.1.0
num2words==0.5.10
oauthlib==3.1.0
openpy==0.0.2
packaging==20.9
parser-libraries==3.6
parser-main==4.2
parser-module==4.8
pdfkit==0.6.1
Pillow==8.1.1
protobuf==3.15.8
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyTelegramBotAPI==3.7.7
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3fs==0.6.0
selenium==3.141.0
shutilwhich==1.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
WebOb==1.8.7
wrapt==1.12.1
xhtml2pdf==0.2.5
yarl==1.6.3
zipp==3.4.0
